chapter 22 swing front-end created

+++ src/gr/aueb/cf/ch16/challenges/challenge2/ITwoDimensional.java
@@ -1,9 +1,8 @@
 package gr.aueb.cf.ch16.challenges.challenge2;
 
-
 public interface ITwoDimensional {
     double getArea();
 
-    long getCircumference();
+    double getCircumference();
 
 }
